Extract getResponseType
